added a temp folder with my code and data for the pacific figure
if we are happy with this, i can move the data to the data folder and
move this up a folder and adjust the paths accordingly. For now ive
just dumped it in here as a block. One question mark is over the
workflow, which doesnt show where the files test2A.csv etc… come from,
which i think is from some of Clive’s original code. I can check this
too.